introduce Flight timer, rename config-parameters
1.) config.h specifies for each profile a flight timer in seconds which
will be count down.
the timer starts when throttle is higher than
cfg_FlightTimeTrigger_ThrottleVal and will pause if it is below that.
When the timer has reached 0 the buzzer will beep for 1 second
10 seconds before that you'll hear a short double beep every second
every minute you'll hear a short single beep, also after the timer has
reached zero.
turn transmitter off and on again to reset the timer
2.) rename all config parameters from g_* to cfg_*