preparations for virtual flight mode switch (VFM)
some FCs like APM are using one channel to handle up to 6 different
flight modes.
VFM is intended to realize such a channel by means of a 3-pos and a
2-pos switch.
instead of putting the signal of the switchtes directly on the channel,
both switches will generate a signal on a channel. in that case channel
layout has to be changed. remove the 2 switches from the layout and use
"M" instead for the mode-switch.
define the different values the virtual mode switch should be able to
handle in cfg_VFMSteps in config.h

- fix in RC-Lib Channel.cpp in case of using rc::Output_None as
m_source.
without that we have random values on such channels
- preparations for virtual flight mode switch:
cfg_VFMSteps defines 6 values from -256 to +256
- removing T5X_USE_LESS_ROM_FOR_CHANNEL_ORDER_EVALUATION preprocessor
define as well as code snippets.
- adding parameter check, i.e. assignment is done only if character is
found in channel layout. without that check we might result in invalid
index because -1 is returned.

works fine basically.

TBD:
- some more testing required.
- rethink the way how the channel layout is parsed, since only the first
occurence of each character is checked.
- extend DR/EXPO from 3 to 6 flight modes. FM 0-2 will be used always by
means of the selected flight mode switch. FM 3-5 will be used only in
conjunction with VFM-switch.