introducing setup mode
a first simple start for a setup mode, if connecting USB-cable without
turning on the transmitter, it will automatically turn into so-called
"OperatingMode_Setup".
This currently in a first version just simply sends the values of the
analog inputs to the serial. this can be used for setting
cfg_AnalogSettings in config.h
in the future it is planned to add more functionality to the setup-mode
to configure the T5x via serial and store configuration to EEPROM rather
than hardcoding the config in config.h and compile it to the ROM.